Lucid.Alpine

Documentation

xData_ :: Maybe Text -> Attribute

x-data Declare a new Alpine component and its data for a block of HTML

xBind_

Arguments

:: Text
Attribute name

-> Text

-> Attribute

x-bind Dynamically set HTML attributes on an element

xOn_

Arguments

:: Text
Event name

-> Text

-> Attribute

x-on Listen for browser events on an element

xText_ :: Text -> Attribute

x-text Set the text content of an element

xHtml_ :: Text -> Attribute

x-html Set the inner HTML of an element

xModel_

Arguments

:: [Text]
List of x-model modifiers

-> Text

-> Attribute

x-model Synchronize a piece of data with an input element

xShow_ :: Text -> Attribute

x-show Toggle the visibility of an element

xTransition_

Arguments

:: Maybe Text
Transition directive

-> [Text]
List of x-transition modifiers

-> Text

-> Attribute

x-transition Transition an element in and out using CSS transitions

xFor_ :: Text -> Attribute

x-for Repeat a block of HTML based on a data set

xForKey_ :: Text -> Attribute

xIf_ :: Text -> Attribute

x-if Conditionally add/remove a block of HTML from the page entirely.

xInit_ :: Text -> Attribute

x-init Run code when an element is initialized by Alpine

xEffect_ :: Text -> Attribute

x-effect Execute a script each time one of its dependancies change

xRef_ :: Text -> Attribute

x-ref Reference elements directly by their specified keys using the $refs magic property

xCloak_ :: Attribute

x-cloak Hide a block of HTML until after Alpine is finished initializing its contents

xIgnore_ :: Attribute

x-ignore Prevent a block of HTML from being initialized by Alpine